+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Changes
 
 ### 0.1.1
-- Changing camera with plyer camera on android.
+- Using our own customization of kivy camera module (mainly for android).
 
 ### 0.1.0
 - Adding a `Secure Vault` feature to store secret files.